requirements.txt
numpy
pandas
torch
pillow
torchvision
torchmetrics
tqdm
tensorboard
matplotlib
